microservices/org-service/requirements.txt
Django>=5.0,<6.0
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
django-cors-headers
redis
django-redis
django-filter
python-dotenv
gunicorn
whitenoise
django-phonenumber-field[phonenumbers]
pika
Pillow
requests
ams-shared @ file:///app/ams-shared
